Insertion of arrays
Insert operation in Array or inserting elements in an Array simply
means, Adding one or more elements at specified index/indices to the
existing elements of an array.
It’s only because, When you declare and initializes the values
(elements) of an array, all the elements occupies its memory space
according to the sequence (order) in which they are initialized.

And, By increasing the size of the array by one you basically creates a
room (space) for the new incoming element which you wish to insert at a
specified position in the given array.

Deletion of arrays

Deleting an array is similar to assigning undefined to that particular
element .
Delete the element to be deleted and move the other elements to the left
to keep all elements together.
If array is static memory allocated, memory cannot be deleted. Memory is
released when program terminates.
If array is dynamic memory allocated, extra memory can be deleted by
using realloc.
The time complexity to delete an element is O(N) time.